Back end logic for bidding and image usage

+++ Back-end/src/main/java/com/internship/AuctionApp/Bids/BidResponse.java
@@ -6,13 +6,13 @@ public class BidResponse {
     private float highestBid;
     private int numberOfBids;
 
-    public BidResponse(Long latest_bidder_id, int highestBid, int numberOfBids) {
+    public BidResponse(Long latest_bidder_id, float highestBid, int numberOfBids) {
         this.latest_bidder_id = latest_bidder_id;
         this.highestBid = highestBid;
         this.numberOfBids = numberOfBids;
     }
 
-    public Long getLatest_bidder_id(){
+    public Long getLatest_bidder_id() {
         return latest_bidder_id;
     }
